+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 2.2.6: July 27th, 2022
 * Link to booking page displayed as a well
 * Fix preselect product from URL, when product has choices